BAN-29 move pages to locale
