--- src/app/(with-header)/auth/email-verify-recovery-password/[email]/page.tsx
@@ -1,7 +1,0 @@
-import { EmailVerifyRecoveryPasswordPage } from '@/app/pages/emailVerifyRecoveryPasswordPage/EmailVerifyRecoveryPasswordPage'
-
-export default async function Page({ params }: { params: { email: string } }) {
-  const { email } = await params
-  const decodedEmail = decodeURIComponent(email)
-  return <EmailVerifyRecoveryPasswordPage email={decodedEmail} />
-}
