+++ src/app/[locale]/(with-header)/auth/email-verify-recovery-password/[...slug]/page.tsx
@@ -1,0 +1,5 @@
+import { EmailVerifyRecoveryPasswordPage } from '@/app/pages'
+
+export default function Page() {
+  return <EmailVerifyRecoveryPasswordPage />
+}
