+++ src/app/[locale]/(with-header)/auth/forgot-password/page.tsx
@@ -1,5 +1,5 @@
-import {ForgotPasswordPage, LoginPage} from "@/app/pages";
+import { ForgotPasswordPage } from '@/app/pages'
 
 export default function Page() {
-    return    <ForgotPasswordPage/>
+  return <ForgotPasswordPage />
 }
