+++ src/app/[locale]/(with-header)/auth/restore-password/[ ... slug]/page.tsx
