+++ src/app/pages/forgotPasswordPage/ForgotPasswordPage.tsx
@@ -1,4 +1,6 @@
-import ForgotPassword from '@features/auth/forgotPassword/ui/ForgotPassword'
+'use client'
+
+import { ForgotPassword } from '@/features'
 
 export const ForgotPasswordPage = () => {
   return <ForgotPassword />
