+++ src/app/pages/restorePassword/RestorePasswordPage.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import NewPasswordForm from '@features/auth/new-password-form/ui/NewPasswordForm'
+import { NewPasswordForm } from '@/features'
 
 export const RestorePasswordPage = ({ email }: { email: string }) => {
   return <NewPasswordForm email={email} />
