+++ src/features/auth/email-verify-recovery-password/ui/EmailVerifyRecoveryPassword.tsx
@@ -7,10 +7,9 @@ import { Button } from '@shared/ui'
 import Image from 'next/image'
 import ImageEmailVerification from '@/assets/auth/email_verification.webp'
 import { useForm } from 'react-hook-form'
-import { withGuestOnly } from '@shared/lib/hoc/withGuestOnly'
 import { useRouter } from 'next/navigation'
 
-const EmailVerifyRecoveryPassword = ({ email }: { email: string }) => {
+const EmailVerifyRecoveryPassword = () => {
   const router = useRouter()
 
   const { handleSubmit, reset } = useForm<FormData>({
@@ -54,4 +53,4 @@ const EmailVerifyRecoveryPassword = ({ email }: { email: string }) => {
   )
 }
 
-export default withGuestOnly(EmailVerifyRecoveryPassword)
+export default EmailVerifyRecoveryPassword
