+++ src/features/index.ts
@@ -1,4 +1,7 @@
 export { default as SignUp } from './auth/signUp/ui/signUp/SignUp'
+export { default as ForgotPassword } from './auth/forgotPassword/ui/ForgotPassword'
+export { default as NewPasswordForm } from './auth/new-password-form/ui/NewPasswordForm'
+export { default as EmailVerifyRecoveryPassword } from './auth/email-verify-recovery-password/ui/EmailVerifyRecoveryPassword'
 
 export { LinkPage } from './auth/components'
 export { PresentationPage } from './auth/components'
